+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.40.5](https://github.com/lyne-design-system/lyne-components/compare/v0.40.4...v0.40.5) (2023-09-05)
+
+### Features
+
+- **sbb-form-field-clear:** initial implementation ([#1962](https://github.com/lyne-design-system/lyne-components/issues/1962)) ([2075dfa](https://github.com/lyne-design-system/lyne-components/commit/2075dfa13f5c05aa5654229fc292d55e58009443))
+
+### [0.40.4](https://github.com/lyne-design-system/lyne-components/compare/v0.40.3...v0.40.4) (2023-09-05)
+
+### Bug Fixes
+
+- **sbb-dialog:** react to slot changes ([#1969](https://github.com/lyne-design-system/lyne-components/issues/1969)) ([326e208](https://github.com/lyne-design-system/lyne-components/commit/326e20864036da6084cc2f980af531de2c53640f))
+
 ### [0.40.3](https://github.com/lyne-design-system/lyne-components/compare/v0.40.2...v0.40.3) (2023-09-05)
 
 ### Features